refactor: extract NowPlaying to separate class
